fix targets
